Refactor middleware logic, Make invoice amount configurable
