+++ ln/lnd.go
@@ -23,10 +23,6 @@ type LNDoptions struct {
 	MacaroonHex  string
 }
 
-type LNClient interface {
-	AddInvoice(ctx context.Context, req *lnrpc.Invoice, options ...grpc.CallOption) (*lnrpc.AddInvoiceResponse, error)
-}
-
 type LNDWrapper struct {
 	client lnrpc.LightningClient
 }
